Conflicts:
	code/DriveIT/DriveIT.Models/CarDto.cs
	code/DriveIT/DriveIT.WindowsClient/ViewModels/CarViewModel.cs

+++ code/DriveIT/DriveIT.Entities/Car.cs
@@ -20,7 +20,7 @@ namespace DriveIT.Entities
         public string Fuel { get; set; }
         public string Drive { get; set; }
         public string Transmission { get; set; }
-        public int TopSpeed { get; set; }
+        public float TopSpeed { get; set; }
         public float NoughtTo100 { get; set; }
         public List<string> ImagePaths { get; set; }
     }
